Add .bash_profile shell script and documentation
Document the purpose of the .bash_profile and .bashrc files.
Now the conditions for loading both files are clear.

The .bash_profile script is sourced by login shells, interactive or not.
On the other hand, the manual says the .bashrc script
is sourced by interactive non-login shells only.
What about interactive login shells then?
If the login shell is interactive, then sourcing .bashrc is beneficial.
If not, just .bash_profile will suffice.