Swap order of the link function's arguments
It is more natural to write "link file to target" than "link to target, file".
The order of the link function's arguments should reflect that.
Existing uses of the link function in recipes have also been updated
due to this change. Now the order of the arguments also matches
the structure of the rules.