Define basic configuration target
Not all configuration files are necessary on all systems.
The new basic phony target configures fundamental software such as:

 - Shells
 - Git
 - Editors

The basic target replaces all as the new default target.
This prevents accidental linking of irrelevant configuration files
on systems that have no use for them. For example:
Sublime Text configuration files do not need to be linked on Termux.