.bashrc: indicate empty repositories in git prompt
Resolving the HEAD of an empty repository with git rev-parse
returns an error and produces empty output since there are no commits.
Indicate that the repository is empty with the empty set symbol.

It makes no sense to compute branch information in this case either,
so do it only when the repository is not empty.