.bashrc: refactor terminal formatting functions
While the flexibility to write an arbitrary sequence of strings
and escape codes to the terminal is useful, the most common use case
consists of uniformly formatting a single block of text and resetting.
For that reason, the formatting functions have been split into two:

 - write: processes an arbitrary sequence of escape codes and strings
 - format: applies formatting to the first argument and resets

This dramatically simplifies the prompt building code
and makes .bashrc pass shellcheck.