.bashrc: receive the status code as an argument
The prompt-error-code function now takes the status code of the last
command as its sole argument, allowing the prompt building code more
flexibility when using the function and positioning the status code
in the prompt. It should now be possible to use it anywhere in the
prompt regardless of any other commands that may run during the process.