.bashrc: use PROMPT_COMMAND to build the prompt
The prompt-command function is run every time the shell writes prompts.
It allows more fine grained control over the prompt generation process.
This is exemplified by the code for saving the status code of the last
program in a variable and then passing it to the prompt-error-code
function, allowing it to be placed anywhere in the prompt.