.bashrc: support escape modes in terminal-write
For a moment I actually thought I was going to get away with this
relatively simple function that just turns its arguments into
terminal escape codes. Unfortunately it's not enough.
This thing is just gonna have to do a lot more.

The terminal-write function had to become more complicated because of
the need to escape non-printing characters for the shell and readline.
If this isn't done, they won't be able to tell where the cursor is
and will end up overwriting parts of the screen when they write to it.

So now terminal-write has more state: the current escaping mode.
If the caller passes as argument escape=bash or escape=readline,
everything after that directive will be escaped as specified.
It can be reset with escape=''.