.bashrc: use constant-width status code in prompt
It's annoying when the prompt changes shape all the time due to prompt
elements popping in and out of existence. The error code, for example,
would randomly shift the position of the next element by a couple of
characters whenever a command exited unsuccessfully.

So now the status code is always shown and it is padded to three digits
since process status codes range from 0 to 255.