.vimrc: don't strip trailing spaces on diff files
Make it so trailing spaces are only stripped if it's not a diff file.
Git diffs and patches have lines with nothing but whitespace in them.
Those spaces have to be there but vim was automatically deleting them.
This has caused so many patches to fail to apply it's not even funny.