bin: add useless alpha channel deleter script
Deletes useless alpha channels from RGBA images.
The alpha channel is useless when it's opaque
throughout the entire image. In those cases
there is no transparency at all and there is
no point in storing the alpha channel.

Overwrites the original image but that's okay
since it's actually lossless operation.