gpg: delete reference to Riseup PGP best practices
There's a notice on it that says it was written for
GNU Privacy Guard version 1.4. Although progress has
been made towards reviewing it, in the end the issue
was closed with the following message:

> We are killing this document as GnuPG's work has been doing
> good defaults for a while and the guide is by itself problematic.

So there's too much legacy stuff in this configuration
and it won't be getting updated. Time to do it again.