gpg: delete CAST5 algorithm from preferences
GNU Privacy Guard documentation refers to it as an "old" algorithm
which uses 64 bit blocks as opposed to modern algorithms which use
128 bit blocks instead and are presumably stronger. Cursory search
of the internet does not reveal any obvious problems with CAST5
but let's drop it from the preferences list anyway due to what the
documentation says. AES seems to be the gold standard anyway.