gpg: configure defaults for quick key generation
Default to creating a certification only master key and a subkey
and a subkey for digital signatures, both using elliptic curve
cryptography with the Ed25519 curve as is accepted best practice.

Signing subkeys are generated instead of encryption subkeys
because signing keys are expendable unlike encryption keys
which may require special provisions such as printing out
to paper for backup purposes.

This is because the compromise of a signing key is not really
a big deal, simple key revocation and rotation will suffice
to get things back on track. The key can even be deleted
afterwards with no ill effects.

The same is not true for encryption keys: compromises make it
ill advised to continue their use but it is not safe to delete
them entirely, they must still be kept around in order to decrypt
old data. Deleting or losing the key can cause encrypted data to
be lost for all time. Therefore, copies of important encryption
keys should be made, printed out to paper and stored in a safe
location just in case they're ever needed again.

As a result of the complexity of these requirements and best practices,
generating encryption keys is a more involved process that should be
done deliberately by the user, not by default. Therefore, the generation
of encryption subkeys is disabled by default in quick key generation.