gpg: configure the default keyserver URL
Default to keys.openpgp.org which is a modern verifying keyserver.
Looks like this is going to be the standard keyserver from now on.

It verifies that users who upload keys actually
control the emails listed in their user IDs,
providing a bare minimum of trustworthiness
to the keys downloaded from there.

It also protects people's privacy much better
than the old synchronizing keyservers which
were apparently incapable of deleting data
which was obviously not GDPR compliant.